reorganize and refactor user home path checks
